+++ crates/wdl-engine/src/backend/lsf_apptainer.rs
@@ -46,6 +46,7 @@ use crate::config::Config;
 use crate::config::LsfApptainerBackendConfig;
 use crate::config::TaskResourceLimitBehavior;
 use crate::v1;
+use crate::v1::ContainerSource;
 
 /// The name of the file where the Apptainer command invocation will be written.
 const APPTAINER_COMMAND_FILE_NAME: &str = "apptainer_command";
@@ -71,7 +72,7 @@ struct LsfApptainerTaskRequest {
     /// The task spawn request.
     spawn_request: TaskSpawnRequest,
     /// The requested container for the task.
-    container: String,
+    container: ContainerSource,
     /// The requested CPU reservation for the task.
     required_cpu: f64,
     /// The requested memory reservation for the task.
@@ -465,11 +466,13 @@ impl TaskExecutionBackend for LsfApptainerBackend {
                 }
             }
         }
+        let container = v1::container(requirements, self.engine_config.task.container.as_deref());
+        if let ContainerSource::Unknown(_) = &container {
+            bail!("LSF Apptainer backend does not support unknown container source `{container:#}`")
+        }
+
         Ok(super::TaskExecutionConstraints {
-            container: Some(
-                v1::container(requirements, self.engine_config.task.container.as_deref())
-                    .into_owned(),
-            ),
+            container: Some(format!("{container:#}")),
             cpu: required_cpu,
             memory: required_memory.as_u64().try_into().unwrap_or(i64::MAX),
             // TODO ACF 2025-10-16: these are almost certainly wrong
@@ -497,8 +500,10 @@ impl TaskExecutionBackend for LsfApptainerBackend {
         let requirements = request.requirements();
         let hints = request.hints();
 
-        let container =
-            v1::container(requirements, self.engine_config.task.container.as_deref()).into_owned();
+        let container = v1::container(requirements, self.engine_config.task.container.as_deref());
+        if let ContainerSource::Unknown(_) = &container {
+            bail!("LSF Apptainer backend does not support unknown container source `{container:#}`")
+        }
 
         let mut required_cpu = v1::cpu(requirements);
         let mut required_memory = ByteSize::b(v1::memory(requirements)? as u64);